+++ Demo/Demo/Keyboard/DemoKeyboardActionHandler.swift
@@ -32,7 +32,7 @@ class DemoKeyboardActionHandler: StandardKeyboardActionHandler {
         let standard = super.action(for: gesture, on: action)
         switch gesture {
         case .longPress: return longPressAction(for: action) ?? standard
-        case .tap: return tapAction(for: action) ?? standard
+        case .release: return tapAction(for: action) ?? standard
         default: return standard
         }
     }