+++ Demo/ProDemo/Keyboard/DemoKeyboardActionHandler.swift
@@ -17,8 +17,8 @@ import UIKit
  register and use a custom keyboard action handler.
 
  If you change the ``DemoKeyboardLayoutProvider`` to show an
- `.image` action, this class will copy the images on tap and
- save them on long press.
+ `.image` action, this class will copy the images on release
+ and save them on long press.
  */
 class DemoKeyboardActionHandler: StandardKeyboardActionHandler {
 
@@ -32,7 +32,7 @@ class DemoKeyboardActionHandler: StandardKeyboardActionHandler {
         let standard = super.action(for: gesture, on: action)
         switch gesture {
         case .longPress: return longPressAction(for: action) ?? standard
-        case .tap: return tapAction(for: action) ?? standard
+        case .release: return releaseAction(for: action) ?? standard
         default: return standard
         }
     }
@@ -47,7 +47,7 @@ class DemoKeyboardActionHandler: StandardKeyboardActionHandler {
         }
     }
 
-    func tapAction(for action: KeyboardAction) -> KeyboardAction.GestureAction? {
+    func releaseAction(for action: KeyboardAction) -> KeyboardAction.GestureAction? {
         switch action {
         case .image(_, _, let imageName): return { [weak self] _ in self?.copyImage(named: imageName) }
         default: return nil