+++ Sources/KeyboardKit/Autocomplete/AutocompleteToolbar.swift
@@ -222,7 +222,7 @@ public extension AutocompleteToolbar {
         let proxy = controller.textDocumentProxy
         let actionHandler = controller.keyboardActionHandler
         proxy.insertAutocompleteSuggestion(suggestion)
-        actionHandler.handle(.tap, on: .character(""))
+        actionHandler.handle(.release, on: .character(""))
     }
 }