+++ Sources/KeyboardKit/Callouts/ActionCalloutContext.swift
@@ -144,7 +144,7 @@ open class ActionCalloutContext: ObservableObject {
      */
     open func handleSelectedAction() {
         guard let action = selectedAction else { return }
-        actionHandler.handle(.tap, on: action)
+        actionHandler.handle(.release, on: action)
     }
     
     /**